backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
tzdata==2025.2
